First version of scaffolding code.  Closes seven5/seven5#9.
Improved some error handling inside the m2connector to warn us about getting strange messages from the mongrel2 server.  It is sporadic, but sometimes we appear to get messages with URI like @* that seem to indicate a json service!?!  Tried to brace the code against this.

Added scaffolding.go which implements ScaffoldRestService, the entry point to using scaffolding by users.

Fixed a HORRIBLE bug in the gobstore where deleting the last item from an index caused the gob system to panic.

Moved BACK to bradfitz's implementation of memcache interface after he pushed a change to allow us to work with 2012.01.20 weekly.