Split the session manager out.
Previously, the session manager was entangled with the cookie manager.  Both of these interfaces/impls
were simplified by splitting the two apart.  Now the cookie manager _just_ handles cookies and knows
nothing about what they are used for.  SessionManager continues to work the same way, but now is
independent.  The independent SessionManager meant that Dispatchers who want to deal with sessions
have to hold a reference to one.

Preparation for a new kind of a session manager that understands users.