Updated the seven5 infrastructure to support a "basic" user type.
The new building block is user/basic.go which implements a user resource with a little help from the
application.  It required refactoring the simple session object so it was easier to re-use without having
to mess with the internals.  It also required taking out the Fetcher type and requiring applications that
want to "fetch" from a connector handle that themselves out of band.  This was required because of import
loop and frankly is probably a better design.  Improved the DeploymentEnvironment object so there is now
a clean difference between DeploymentEnvirnoment and EnvironmentVars.  Added a Heroku DeploymentEnvironment
to handle that common case.

BaseDispatcher now has to be able to take a session manager when it's created.  This is a lot gunkier than
I would like.