Updated seven5 to auto-recompile dart web components.
Code generation is now done once at startup since it can't change during runtime and the static files are needed by the dart2js compile.

Added code to detect cases when we need to recompute the dart web components and when we need to recompute the JS.

Further scaled back the project because now the client app calls the code generation in main.

Added support for controlling capitalization of the names in the generated Dart code.