Improve IndexOnlyComponent to send all requests to the singular.
Prior to this change, there was some leftover code in both IndexOnlyComponent
and SimpleIdComponent remaining from when these were part of the tutorial.
These two types were "extracted" from the tutorial and into seven5 as part
of the tutorial development, and it was done badly.  Once the leftovers were
removed, the tests in the tutorial broke because they depended on the previous
code.  I added the machinery to IndexOnlyComponent to send _all_ requests
for /plural/foo to /singular/foo, not just /plural/index.html.  This is
usually what you want, because you want all the html files together in the
singular directory, even if you show the user /plural/index.html.